cafe_project/requirements.txt
Django==5.1.6
django-browser-reload
django-tailwind
pillow==10.2.0
qrcode[pil]
psycopg2-binary
python-dotenv
